Streamline setup() and remove orphan char w/o flicker.
In setup(), use pinmode type INPUT_PULLUP. It does the same thing as calling digitalwrite() on an input pin, but is more descriptive of the intent. The type INPUT_PULLUP was introduced in Arduino IDE v1+, so this will break older IDE installations.

To allow removing orphaned characters in the counter without the display flickering, I implemented a boolean flag called "changed". Every time the count is changed from a button press I set the flag to true. Then I only change the display if the flag is set, and immediately clear the flag.

I haven't compiled these changes so I may have introduced a typo. My apologies in advance if I have.